requirements.txt
elasticsearch >= 8.12.1
elasticsearch-async >= 6.2.0
load_dotenv >= 0.1.0
tqdm >= 4.66.2
urllib3 >= 2.2.1
pandas
torch
transformers
sentence-transformers